backend/requirements.txt
fastapi
uvicorn
pillow
openai
python-dotenv
requests